uf2conv: add more searchpaths, fix pathcheck
Some distributions like Arch Linux still use '/mnt' as the
primary mounting directory.
If the 'uf2conv.py' script was called via sudo, it did not
find any devices because it searched at '/media/root', which
usually is unused.
Therefore the ${SUDO_USER} env-variable is evaluated as a
searchpath, if it exists.